+++ Examples/SwiftUI/Superwall-SwiftUI/Services/SuperwallService.swift
@@ -29,10 +29,10 @@ final class SuperwallService {
     // }
 
     Superwall.configure(
-      apiKey: apiKey,
-      delegate: shared/*,
+      apiKey: apiKey/*,
       purchaseController: shared*/
     )
+    Superwall.shared.delegate = shared
 
     // Getting our logged in status from Superwall.
     shared.isLoggedIn.send(Superwall.shared.isLoggedIn)
@@ -54,8 +54,8 @@ final class SuperwallService {
     Superwall.shared.handleDeepLink(url)
   }
 
-  static func setName(to name: String) {
-    Superwall.shared.setUserAttributes(["firstName": name])
+  static func setName(to name: String) async {
+    await Superwall.shared.setUserAttributes(["firstName": name])
   }
 }