+++ Examples/SwiftUI/Superwall-SwiftUI/WelcomeView.swift
@@ -66,7 +66,7 @@ struct WelcomeView: View {
   private func logInButton() -> some View {
     BrandedButton(title: "Log In") {
       Task {
-        SuperwallService.setName(to: name)
+        await SuperwallService.setName(to: name)
         await SuperwallService.identify()
         isLoggedIn = true
       }