+++ Examples/UIKit-Swift/Superwall-UIKit-Swift/WelcomeViewController.swift
@@ -39,7 +39,7 @@ final class WelcomeViewController: UIViewController {
     UserDefaults.standard.setValue(true, forKey: "IsLoggedIn")
     Task {
       if let name = textField.text {
-        SuperwallService.setName(to: name)
+        await SuperwallService.setName(to: name)
       }
       await SuperwallService.identify()
       next()