+++ Sources/SuperwallKit/Dependencies/DependencyContainer.swift
@@ -15,8 +15,6 @@ import Combine
 ///
 /// Objects only need `unowned` references to the dependencies injected into them because
 /// `DependencyContainer` is owned by the `Superwall` class.
-///
-/// Idea taken from: [swiftbysundell.com](https://www.swiftbysundell.com/articles/dependency-injection-using-factories-in-swift/)
 final class DependencyContainer {
   // swiftlint:disable implicitly_unwrapped_optional
   var configManager: ConfigManager!
@@ -39,16 +37,12 @@ final class DependencyContainer {
   // swiftlint:enable implicitly_unwrapped_optional
 
   init(
-    swiftDelegate: SuperwallDelegate? = nil,
-    objcDelegate: SuperwallDelegateObjc? = nil,
     swiftPurchaseController: PurchaseController? = nil,
     objcPurchaseController: PurchaseControllerObjc? = nil,
     options: SuperwallOptions? = nil
   ) {
     storeKitManager = StoreKitManager(factory: self)
     delegateAdapter = SuperwallDelegateAdapter(
-      swiftDelegate: swiftDelegate,
-      objcDelegate: objcDelegate,
       swiftPurchaseController: swiftPurchaseController,
       objcPurchaseController: objcPurchaseController
     )
@@ -267,9 +261,11 @@ extension DependencyContainer: RequestFactory {
 extension DependencyContainer: ApiFactory {
   func makeHeaders(
     fromRequest request: URLRequest,
+    isForDebugging: Bool,
     requestId: String
   ) -> [String: String] {
-    let auth = "Bearer \(storage.apiKey)"
+    let key = isForDebugging ? storage.debugKey : storage.apiKey
+    let auth = "Bearer \(key)"
     let headers = [
       "Authorization": auth,
       "X-Platform": "iOS",