+++ Sources/SuperwallKit/Graveyard/SuperwallGraveyard.swift
@@ -10,7 +10,7 @@ import UIKit
 public extension Superwall {
   // MARK: - Unavailable methods
   // TODO: Fix deprecation here
-  /*@available(*, unavailable, renamed: "configure(apiKey:delegate:options:completion:)")
+  /*@available(*, unavailable, renamed: "configure(apiKey:purchaseController:options:completion:)")
   @discardableResult
   @objc static func configure(
     apiKey: String,